requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==2.1.2
django-filebrowser-no-grappelli==3.7.5
django-tinymce4-lite==1.7.3
idna==2.7
jsmin==2.2.2
mysqlclient==1.3.13
Pillow==5.2.0
pkg-resources==0.0.0
pytz==2018.5
requests==2.19.1
urllib3==1.23
